release

+++ musixise-box-shop/src/main/kotlin/com/musixise/musixisebox/shop/enums/ProductTypeEnum.kt
@@ -1,6 +1,6 @@
 package com.musixise.musixisebox.shop.enums
 
-enum class ProductTypeEnum(val type: Int) {
+enum class ProductTypeEnum(val type: Long) {
 
     MUSIX_BOX(1),
     MUSIX_DOWNLOAD(100);
